+++ framework/core/Console/Files/MakeRequest.php
@@ -9,7 +9,7 @@
 
 declare(strict_types=1);
 
-namespace Brocooly\Console;
+namespace Brocooly\Console\Files;
 
 use Brocooly\Request\Request;
 use Symfony\Component\Console\Input\InputArgument;