+++ framework/core/Console/Support/ClearCache.php
@@ -9,8 +9,9 @@
 
 declare(strict_types=1);
 
-namespace Brocooly\Console;
+namespace Brocooly\Console\Support;
 
+use Brocooly\Console\SymfonyStyleTrait;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;